order_service/api/Suppliers/serializers.py
from rest_framework import serializers
from orders.models import Suppliers as Supplier

class PublicSupplierSerializer(serializers.ModelSerializer):
    class Meta:
        model = Supplier
        fields = ['id', 'name', 'performance_rating']

class AdminSupplierSerializer(serializers.ModelSerializer):
    class Meta:
        model = Supplier
        fields = '__all__'
