order_service/requirements.txt
django 
djangorestframework
psycopg2-binary 
pika 
python-dotenv 
celery 
redis 
requests
django-celery-results
django-celery-beat
django-cors-headers
djangorestframework-simplejwt